requirements.txt
# Core ML/DL Dependencies
tensorflow>=2.10.0
numpy>=1.23.0

# Data Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Machine Learning Utilities
scikit-learn>=1.2.0

# Image Processing
Pillow>=9.0.0

# Web Application
streamlit>=1.20.0
